+++ packages/svelte-query/tests/useMutationState/useMutationState.svelte.test.ts
@@ -1,9 +1,9 @@
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { fireEvent, render } from '@testing-library/svelte'
 import { QueryClient } from '@tanstack/query-core'
 import { queryKey, sleep } from '@tanstack/query-test-utils'
-import BaseExample from './BaseExample.svelte'
-import SelectExample from './SelectExample.svelte'
+import Base from './Base.svelte'
+import Select from './Select.svelte'
 import type { Mutation } from '@tanstack/query-core'
 
 describe('useMutationState', () => {
@@ -19,7 +19,7 @@ describe('useMutationState', () => {
     vi.useRealTimers()
   })
 
-  test('should run few mutation functions and check from useMutationState', async () => {
+  it('should run few mutation functions and check from useMutationState', async () => {
     const successKey = queryKey()
     const errorKey = queryKey()
     const successMutationFn = vi.fn(() => sleep(10).then(() => 'data'))
@@ -29,7 +29,7 @@ describe('useMutationState', () => {
         sleep(20).then(() => Promise.reject(new Error('error'))),
       )
 
-    const rendered = render(BaseExample, {
+    const rendered = render(Base, {
       props: {
         queryClient,
         successMutationOpts: () => ({
@@ -54,7 +54,7 @@ describe('useMutationState', () => {
     expect(rendered.getByText('Data: ["success","error"]')).toBeInTheDocument()
   })
 
-  test('should select specific type of mutation ( i.e: error only )', async () => {
+  it('should select specific type of mutation ( i.e: error only )', async () => {
     const successKey = queryKey()
     const errorKey = queryKey()
     const successMutationFn = vi.fn(() => sleep(10).then(() => 'data'))
@@ -64,7 +64,7 @@ describe('useMutationState', () => {
         sleep(20).then(() => Promise.reject(new Error('error'))),
       )
 
-    const rendered = render(BaseExample, {
+    const rendered = render(Base, {
       props: {
         queryClient,
         successMutationOpts: () => ({
@@ -92,10 +92,10 @@ describe('useMutationState', () => {
     expect(rendered.getByText('Data: ["error"]')).toBeInTheDocument()
   })
 
-  test('should return selected value when using select option', async () => {
+  it('should return selected value when using select option', async () => {
     const mutationKey = queryKey()
 
-    const rendered = render(SelectExample, {
+    const rendered = render(Select, {
       props: {
         queryClient,
         mutationOpts: () => ({
@@ -119,7 +119,7 @@ describe('useMutationState', () => {
     expect(rendered.getByText('Variables: ["success"]')).toBeInTheDocument()
   })
 
-  test('should select specific mutation using mutation key', async () => {
+  it('should select specific mutation using mutation key', async () => {
     const successKey = queryKey()
     const errorKey = queryKey()
     const successMutationFn = vi.fn(() => sleep(10).then(() => 'data'))
@@ -129,7 +129,7 @@ describe('useMutationState', () => {
         sleep(20).then(() => Promise.reject(new Error('error'))),
       )
 
-    const rendered = render(BaseExample, {
+    const rendered = render(Base, {
       props: {
         queryClient,
         successMutationOpts: () => ({